Initial crack at scoped association preloading
The idea is that when you call a named scope on an association
collection, we'll preload that named scope for that collection
for all of the instances in your originating collection.

Currently a very janky implementation -- basically creates a new
read-only association in the namespace of the receiving class each time
a distinct scope is used. Need to find a way to clean that up.