Less janky, but still janky, scoped preloading
Refactored the scoped preloading code so that there is a single "dummy"
association that is always used for scoped preloading. Eager record
manually creates an association collection with the correct scoping
information and adds it to the records in question before telling AR to
preload the association (by passing the reflection directly), so we're
no longer creating a new association for each set of conditions that
might be passed in. But the code is still very messy and needs cleanup
-- at least this is a general approach I like.