Group originating collection by class
With STI classes, we may load instances of multiple subclasses in a
single query. Some of these subclasses may define associations that
others do not. If you attempt to preload such an association across the
entire collection, you'll get an error that the association doesn't
exist for some of the records.

The solution is to group the originating collection by the record's
actual class, thus only eagerly preloading objects of the same class as
the record in question.